Split IndexationStrategy tests into ES5 and ES8 variants
ES5 tests moved to ElasticSearch5/ subfolder and implement the
ElasticSearch5Test marker interface. ES8 tests stay in place,
drop the _type field and include @type in the document body.
